push bt module 4

+++ Module 4/ss7_Spring_Data_Repository/Thuc_hanh/quan_ly_khach_hang_quan_ly_tinh/src/main/java/com/example/quan_ly_khach_hang_quan_ly_tinh/repository/IProviceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IProviceRepository extends JpaRepository<Province,Integer> {
+public interface IProviceRepository extends JpaRepository<Province,Long> {
 }
